backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
redis==5.0.1
google-generativeai==0.3.2
numpy==1.26.3
PyPDF2==3.0.1
python-docx==1.1.0
requests==2.31.0
aiofiles==23.2.1
pgvector==0.2.4
pydantic[email]==2.5.3
email-validator>=2.3.0
